contributors/gadams/incoming/cssom/cssmediarule-interface-object.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD//XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>CSSOM - Interface Object - CSSMediaRule</title>
    <link rel="author" title="Glenn Adams" href="mailto:glenn@skynav.com"/>
    <link rel="help" href="http://www.w3.org/TR/cssom/#the-cssmediarule-interface"/>
    <meta name="flags" content="dom"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
  </head>
  <body>
    <div id="log"></div>
    <div id="box"></div>
    <script><![CDATA[

      test(function(){

        assert_true(CSSMediaRule !== undefined);

      }, 'CSSMediaRule !== undefined');

      test(function(){

        assert_true(CSSMediaRule instanceof Object);

      }, 'CSSMediaRule instanceof Object');

      test(function(){

        assert_true(CSSMediaRule instanceof Function);

      }, 'CSSMediaRule instanceof Function');

      test(function(){

        assert_true(CSSMediaRule instanceof CSSRule);

      }, 'CSSMediaRule instanceof CSSRule');

    ]]>
    </script>
  </body>
</html>
